+++ modules/grunt-prerender/tasks/index.ts
@@ -1,6 +1,6 @@
-import * as Prerender from './src/prerender';
+import * as Prerender from './prerender';
 
-export * from './src/prerender';
+export * from './prerender';
 
 // needs to be written like this otherwise Grunt will fail to load this task
 module.exports = Prerender;
